+++ cli/azd/pkg/project/framework_service_docker.go
@@ -14,6 +14,7 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/async"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
+	"github.com/azure/azure-dev/cli/azd/pkg/input"
 	"github.com/azure/azure-dev/cli/azd/pkg/output"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/docker"
@@ -68,6 +69,7 @@ type dockerProject struct {
 	docker          docker.Docker
 	framework       FrameworkService
 	containerHelper *ContainerHelper
+	console         input.Console
 }
 
 // NewDockerProject creates a new instance of a Azd project that
@@ -76,11 +78,13 @@ func NewDockerProject(
 	env *environment.Environment,
 	docker docker.Docker,
 	containerHelper *ContainerHelper,
+	console input.Console,
 ) CompositeFrameworkService {
 	return &dockerProject{
 		env:             env,
 		docker:          docker,
 		containerHelper: containerHelper,
+		console:         console,
 	}
 }
 
@@ -151,6 +155,13 @@ func (p *dockerProject) Build(
 
 			// Build the container
 			task.SetProgress(NewServiceProgress("Building Docker image"))
+
+			previewerWriter := p.console.ShowPreviewer(ctx,
+				&input.ShowPreviewerOptions{
+					Prefix:       "  ",
+					MaxLineCount: 8,
+					Title:        "Docker Output",
+				})
 			imageId, err := p.docker.Build(
 				ctx,
 				serviceConfig.Path(),
@@ -159,7 +170,9 @@ func (p *dockerProject) Build(
 				dockerOptions.Context,
 				imageName,
 				dockerOptions.BuildArgs,
+				previewerWriter,
 			)
+			p.console.StopPreviewer(ctx)
 			if err != nil {
 				task.SetError(fmt.Errorf("building container: %s at %s: %w", serviceConfig.Name, dockerOptions.Context, err))
 				return