+++ templates/common/infra/terraform/core/host/appservice/appservicenode/appservicenode.tf
@@ -62,3 +62,14 @@ resource "azurerm_linux_web_app" "web" {
     }
   }
 }
+
+# This is a temporary solution until the azurerm provider supports the basicPublishingCredentialsPolicies resource type
+resource "null_resource" "webapp_basic_auth_disable" {
+  triggers = {
+    account = azurerm_linux_web_app.web.name
+  }
+
+  provisioner "local-exec" {
+    command = "az resource update --resource-group ${var.rg_name} --name ftp --namespace Microsoft.Web --resource-type basicPublishingCredentialsPolicies --parent sites/${azurerm_linux_web_app.web.name} --set properties.allow=false && az resource update --resource-group ${var.rg_name} --name scm --namespace Microsoft.Web --resource-type basicPublishingCredentialsPolicies --parent sites/${azurerm_linux_web_app.web.name} --set properties.allow=false"
+  }
+}